requirements.txt
# --- Core ---
python-dotenv>=1.0.0

# --- LLM / Embeddings ---
ollama>=0.3.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-ollama>=0.1.0
langchain-text-splitters>=0.2.0
# Cloud backends (HF Spaces)
groq>=0.9.0
sentence-transformers>=3.0.0

# --- Vector Store ---
lancedb>=0.13.0

# --- Scraping ---
requests>=2.32.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# --- Text Processing ---
nltk>=3.8.0
tiktoken>=0.7.0          # chunk size counting

# --- UI ---
chainlit>=1.2.0

# --- Voice ---
faster-whisper>=1.0.0
edge-tts>=6.1.0          # cross-platform TTS (HF Spaces / Linux)
# piper-tts>=1.2.0       # optional upgrade — needs .onnx model download

# --- MCP (Phase 7) ---
# mcp>=1.0.0
